Rename modules to accountManager, personaManager
