+++ packages/sanes-chrome-extension/src/logic/persona/index.ts
@@ -1,2 +1,2 @@
 export * from './persona';
-export * from './personamanager';
+export * from './personaManager';
